Rename module to prefixed_storage
