+++ packages/storage/src/lib.rs
@@ -1,6 +1,6 @@
 mod bucket;
 mod namespace_helpers;
-mod prefix;
+mod prefixed_storage;
 mod sequence;
 mod singleton;
 mod transactions;
@@ -8,7 +8,7 @@ mod type_helpers;
 mod typed;
 
 pub use bucket::{bucket, bucket_read, Bucket, ReadonlyBucket};
-pub use prefix::{prefixed, prefixed_read, PrefixedStorage, ReadonlyPrefixedStorage};
+pub use prefixed_storage::{prefixed, prefixed_read, PrefixedStorage, ReadonlyPrefixedStorage};
 pub use sequence::{currval, nextval, sequence};
 pub use singleton::{singleton, singleton_read, ReadonlySingleton, Singleton};
 pub use transactions::{transactional, transactional_deps, RepLog, StorageTransaction};
